[ifmap-server]: Do not throw exception when some entries are already removed
The fix avoids raising exception when a non-existed link is being removed from the graph data structure. The reason is the link is removed already and no longer presents in the graph, hence we could ignore safely if there is no link in "Removal" case. The exception causing a problem, it stops the execution of propagating API data changes to control node.

Closes-Bug: #1688227