[ifmap-server]: Fixes for ebay/workday ifmap server issue
The fix avoids raising exception when a non-existed link is being
removed from the graph data structure. The reason is the link is
removed already and no longer presents in the graph, hence we could ignore
safely if there is no link in "Removal" case.
The exception causing a problem, it stops the execution of propagating
API data changes to control node.
This is additional fix added to the previous fix.
https://github.com/Juniper/contrail-third-party-packages/pull/72

Closes-Bug: #1688227